+++ benchmark/Tests/Commands/ClearOfflineDataCommand.swift
@@ -5,8 +5,7 @@ import XCTest
 struct ClearOfflineDataCommand: AsyncCommand, Decodable {
     @MainActor
     func execute(context: Context) async throws {
-        let options = ResourceOptionsManager.default.resourceOptions
-        let manager = OfflineManager(resourceOptions: options)
+        let manager = OfflineManager()
 
         let packs = try await manager.allStylePacks()
 
@@ -14,7 +13,7 @@ struct ClearOfflineDataCommand: AsyncCommand, Decodable {
             manager.removeStylePack(for: StyleURI(rawValue: pack.styleURI)!)
         }
 
-        try await MapboxMap.clearData(for: options)
+        try await MapboxMap.clearData()
 
         let tileStore = TileStore.default
         tileStore.setOptionForKey(TileStoreOptions.diskQuota, value: 0)