+++ mapbox-maps-ios/Apps/StressTest/StressTest/ViewController.swift
@@ -294,9 +294,7 @@ class ViewController: UIViewController {
 
         // Configure the snapshotter object with its default access
         // token, size, map style, and camera.
-        let options = MapSnapshotOptions(size: CGSize(width: 300, height: 300),
-                                         pixelRatio: 1,
-                                         resourceOptions: mapInitOptions.resourceOptions)
+        let options = MapSnapshotOptions(size: CGSize(width: 300, height: 300), pixelRatio: 1)
 
         print("Creating snapshotter")
         let snapshotter = Snapshotter(options: options)